+++ ChangeLog.md
@@ -21,6 +21,7 @@
     - audio.pl list has '--step' option when listing by genre
     - audio.pl list has '--summary-unchecked' option to not prevent display of counters for unchecked properties
     - audio.pl list and normalize verbs have '--album-level' and '--artist-level' options
+    - audio.pl list has '--check-same-year' option
 
 ### 4.19.0